Update atomical-operation-builder.ts
We have fixed the issue where tokens or NFTs with bitworkr set were adding additional outputs in the transaction for storing time and random number information, resulting in increased fees. This caused the fee in the second broadcast to be inconsistent with the first.

In the new version, I have added a judgment for bitworkr to avoid those tokens or NFTs that do not have bitworkr set, which also increases the fee.